Add some desktop-related properties. In particular _NET_CURRENT_DESKTOP (and _NET_NUMBER_OF_DESKTOPS). This is sufficient to work in a basic manner with the ctwm workspace switcher (pager).

+++ workmgr.h
@@ -85,7 +85,7 @@ struct WorkSpaceMgr {
 };
 
 struct WorkSpace {
-  long	              number;
+  int	              number;
   char	              *name;
   char	              *label;
   Image	              *image;
